# nextjs-frontend/app/api/vfs/nodes/[nodeId]/route.ts
"use server";

import { cookies } from "next/headers";
import { NextResponse } from "next/server";

function getApiBaseUrl() {
  return process.env.INTERNAL_API_BASE_URL || process.env.API_BASE_URL || process.env.NEXT_PUBLIC_API_BASE_URL || "http://localhost:8000";
}

export async function DELETE(request: Request, ctx: { params: Promise<{ nodeId: string }> }) {
  const { nodeId } = await ctx.params;
  const cookieStore = await cookies();
  const token = cookieStore.get("accessToken")?.value;
  if (!token) return new NextResponse("unauthorized", { status: 401 });

  const url = new URL(request.url);
  const recursive = url.searchParams.get("recursive") || "";
  const upstreamUrl = new URL(`/api/v1/vfs/nodes/${encodeURIComponent(nodeId)}`, getApiBaseUrl());
  if (recursive) upstreamUrl.searchParams.set("recursive", recursive);

  const upstream = await fetch(upstreamUrl.toString(), {
    method: "DELETE",
    headers: { Authorization: `Bearer ${token}` },
    cache: "no-store",
  });

  const body = await upstream.text();
  return new NextResponse(body || upstream.statusText, {
    status: upstream.status,
    headers: { "content-type": upstream.headers.get("content-type") || "application/json" },
  });
}